multiThread

+++ src/common/util/NeighborComparatorByMBRID.java
@@ -9,7 +9,7 @@ public class NeighborComparatorByMBRID implements Comparator<Neighbor> {
 	@Override
 	public int compare(Neighbor o1, Neighbor o2) {
 		
-		return ((Integer)o1.getMbr().getId()).compareTo((Integer)o2.getMbr().getId());
+		return ((Integer)o1.getMbr().getId()).compareTo(o2.getMbr().getId());
 	}
